requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
requests>=2.31.0
zep-cloud>=3.20.0
graphiti-core>=0.3.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
